+++ lib/physics/circle.dart
@@ -26,11 +26,11 @@ class Circle extends Shape {
       );
 
   @override
-  void draw(Canvas canvas, Color color) {
+  void draw(Canvas canvas, Color color, [bool drawBoundingBox = true]) {
     Shape.paint.style = PaintingStyle.fill;
     Shape.paint.color = color;
     canvas.drawCircle(Offset(position.x, position.y), radius, Shape.paint);
-    super.draw(canvas, color.inverted);
+    if (drawBoundingBox) super.draw(canvas, color.inverted);
   }
 
   Circle copyWith({